Resolve "gérer le fait de créer 2 critiques dans 2 onglets différents"

Closes #101

See merge request igabarra/s3.01-devapp-equipe12!109

+++ symfony/src/Controller/SeriesController.php
@@ -53,7 +53,6 @@ class SeriesController extends AbstractController
             }
             $query = $seriesRepository->findByCriteria($criteria, $search)->getQuery()->getResult();
         } else {
-
             $query = $entityManager->getRepository(Series::class)->queryRandom($seed)->getQuery();
         }
 
@@ -72,7 +71,7 @@ class SeriesController extends AbstractController
         } else {
             $user = $entityManager->getRepository(User::class)
                 ->findOneBy(['email' => $this->getUser()->getUserIdentifier()]);
-        } 
+        }
 
         if ($user) {
             $series_view = $seriesRepository->queryVisionage($user->getId(), $series_id, $seed);
@@ -243,15 +242,18 @@ class SeriesController extends AbstractController
         } else {
             $user = $entityManager->getRepository(User::class)
                 ->findOneBy(['email' => $this->getUser()->getUserIdentifier()]);
+            $rating = $entityManager->getRepository(Rating::class)
+                ->findOneBy(['user' => $user, 'series' => $series]);
+            if ($rating == null) {
+                $rating = new Rating();
+                $rating->setUser($user);
+                $rating->setSeries($series);
+                $rating->setValue($note);
+                $rating->setDate(new \DateTime());
+                $entityManager->persist($rating);
 
-            $rating = new Rating();
-            $rating->setUser($user);
-            $rating->setSeries($series);
-            $rating->setValue($note);
-            $rating->setDate(new \DateTime());
-            $entityManager->persist($rating);
-
-            $entityManager->flush();
+                $entityManager->flush();
+            }
 
             return $this->redirectToRoute('app_series_show', ['id' => $series->getId()], Response::HTTP_SEE_OTHER);
         }
@@ -267,9 +269,11 @@ class SeriesController extends AbstractController
                 ->findOneBy(['email' => $this->getUser()->getUserIdentifier()]);
             $rating = $entityManager->getRepository(Rating::class)
                 ->findOneBy(['user' => $user, 'series' => $series]);
-            $entityManager->remove($rating);
-            $entityManager->flush();
 
+            if ($rating != null) {
+                $entityManager->remove($rating);
+                $entityManager->flush();
+            }
             return $this->redirectToRoute('app_series_show', ['id' => $series->getId()], Response::HTTP_SEE_OTHER);
         }
     }
@@ -290,7 +294,9 @@ class SeriesController extends AbstractController
                 $rating->setSeries($series);
                 $rating->setValue(0);
             }
-            $rating->setComment($request->request->get('comment'));
+            if ($rating->getComment() == null) {
+                $rating->setComment($request->request->get('comment'));
+            }
             $rating->setDate(new \DateTime());
             $entityManager->persist($rating);
             $entityManager->flush();
